Revert "New types uword_t, sword_t"
This reverts commit e26e24ac5b2c744fc5104f7123c5ddcf884dca02.

+++ src/runtime/gc.h
@@ -17,11 +17,10 @@
 #define _GC_H_
 
 #include "sbcl.h"
-#include <stdint.h>
 
 #define PAGE_BYTES BACKEND_PAGE_BYTES
 
-typedef intptr_t page_index_t;
+typedef signed long page_index_t;
 #define PAGE_INDEX_FMT "ld"
 
 typedef signed char generation_index_t;
